pyproject.toml
[tool.poetry]
name = "covid_tracking"
version = "0.1.0"
description = "git@github.com:peksinsara/covid_tracking.git"
authors = ["Sara Peksin <peksin.sara@gmail.com>"]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.4"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.2"
djangorestframework-gis = "^0.16"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
